Extract everything depending on LLBuild into a SwiftDriverExecution library
- Moved llbuild.Swift and the SwiftDriverExecutor into the new library
- Avoid reexporting llbuildSwift through the new library
- Update SPM and CMake builds
