+++ Sources/SwiftDriver/Execution/DriverExecutor.swift
@@ -91,87 +91,3 @@ public protocol JobExecutionDelegate {
   /// Called when a job finished.
   func jobFinished(job: Job, result: ProcessResult, pid: Int)
 }
-
-public final class SwiftDriverExecutor: DriverExecutor {
-  let diagnosticsEngine: DiagnosticsEngine
-  let processSet: ProcessSet
-  let fileSystem: FileSystem
-  public let resolver: ArgsResolver
-  let env: [String: String]
-  
-  public init(diagnosticsEngine: DiagnosticsEngine,
-              processSet: ProcessSet,
-              fileSystem: FileSystem,
-              env: [String: String]) throws {
-    self.diagnosticsEngine = diagnosticsEngine
-    self.processSet = processSet
-    self.fileSystem = fileSystem
-    self.env = env
-    self.resolver = try ArgsResolver(fileSystem: fileSystem)
-  }
-  
-  public func execute(job: Job,
-                      forceResponseFiles: Bool = false,
-                      recordedInputModificationDates: [TypedVirtualPath: Date] = [:]) throws -> ProcessResult {
-    let arguments: [String] = try resolver.resolveArgumentList(for: job,
-                                                               forceResponseFiles: forceResponseFiles)
-    
-    try job.verifyInputsNotModified(since: recordedInputModificationDates,
-                                    fileSystem: fileSystem)
-    
-    if job.requiresInPlaceExecution {
-      for (envVar, value) in job.extraEnvironment {
-        try ProcessEnv.setVar(envVar, value: value)
-      }
-
-      try exec(path: arguments[0], args: arguments)
-      fatalError("unreachable, exec always throws on failure")
-    } else {
-      var childEnv = env
-      childEnv.merge(job.extraEnvironment, uniquingKeysWith: { (_, new) in new })
-
-      let process = try Process.launchProcess(arguments: arguments, env: childEnv)
-      return try process.waitUntilExit()
-    }
-  }
-  
-  public func execute(jobs: [Job],
-                      delegate: JobExecutionDelegate,
-                      numParallelJobs: Int = 1,
-                      forceResponseFiles: Bool = false,
-                      recordedInputModificationDates: [TypedVirtualPath: Date] = [:]
-  ) throws {
-    let llbuildExecutor = MultiJobExecutor(jobs: jobs,
-                                           resolver: resolver,
-                                           executorDelegate: delegate,
-                                           diagnosticsEngine: diagnosticsEngine,
-                                           numParallelJobs: numParallelJobs,
-                                           processSet: processSet,
-                                           forceResponseFiles: forceResponseFiles,
-                                           recordedInputModificationDates: recordedInputModificationDates)
-    try llbuildExecutor.execute(env: env, fileSystem: fileSystem)
-  }
-  
-  @discardableResult
-  public func checkNonZeroExit(args: String..., environment: [String: String] = ProcessEnv.vars) throws -> String {
-    return try Process.checkNonZeroExit(arguments: args, environment: environment)
-  }
-
-  public func description(of job: Job, forceResponseFiles: Bool) throws -> String {
-    let (args, usedResponseFile) = try resolver.resolveArgumentList(for: job, forceResponseFiles: forceResponseFiles)
-    var result = args.joined(separator: " ")
-
-    if usedResponseFile {
-      // Print the response file arguments as a comment.
-      result += " # \(job.commandLine.joinedArguments)"
-    }
-
-    if !job.extraEnvironment.isEmpty {
-      result += " #"
-      for (envVar, val) in job.extraEnvironment {
-        result += " \(envVar)=\(val)"
-      }
-    }
-    return result
-  }
-}
