+++ Sources/SwiftDriverExecution/CMakeLists.txt
@@ -1,0 +1,31 @@
+# This source file is part of the Swift.org open source project
+#
+# Copyright (c) 2020 Apple Inc. and the Swift project authors
+# Licensed under Apache License v2.0 with Runtime Library Exception
+#
+# See http://swift.org/LICENSE.txt for license information
+# See http://swift.org/CONTRIBUTORS.txt for Swift project authors
+
+add_library(SwiftDriverExecution
+  llbuild.swift
+  MultiJobExecutor.swift
+  SwiftDriverExecutor.swift)
+
+target_link_libraries(SwiftDriverExecution PUBLIC
+  TSCBasic
+  TSCUtility
+  SwiftOptions
+  SwiftDriver)
+target_link_libraries(SwiftDriverExecution PRIVATE
+  llbuildSwift)
+
+set_property(GLOBAL APPEND PROPERTY SWIFTDRIVER_EXPORTS SwiftDriverExecution)
+
+# NOTE: workaround for CMake not setting up include flags yet
+set_target_properties(SwiftDriverExecution PROPERTIES
+  INTERFACE_INCLUDE_DIRECTORIES ${CMAKE_Swift_MODULE_DIRECTORY})
+
+install(TARGETS SwiftDriverExecution
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin)
