+++ Sources/swift-driver/main.swift
@@ -9,6 +9,7 @@
 // See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
 //
 //===----------------------------------------------------------------------===//
+import SwiftDriverExecution
 import SwiftDriver
 
 import TSCLibc
